Changed file names to project convention
