+++ client/app/scripts/liveblog-edit/components/polls/poll-component.scss
